Fix issues with jobcontroller class tests
